named control wires correctly and connected them with components
