+++ calyx/src/backend/verilog/mod.rs
@@ -1,1 +1,3 @@
+mod control;
 pub mod gen;
+mod structure;
